pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auras"
version = "0.1.0"
description = "auraS — AI-powered seizure prediction from EEG signals using MindSpore"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Alicja Augustyniak"},
    {name = "Patryk Maciejewski"},
    {name = "Filip Domański"},
]

dependencies = [
    # ── Deep-learning framework ──────────────────────────────────────
    "mindspore>=2.3",
    "mindcv>=0.3",                 # MindSpore CV toolbox (GhostNet, MobileNetV3, ResNet reference)

    # ── EEG / signal processing ──────────────────────────────────────
    "mne>=1.6",                    # EDF loading, filtering, montage handling
    "scipy>=1.11",                 # signal filters, spectral analysis
    "numpy>=1.24,<2",

    # ── Configuration ────────────────────────────────────────────────
    "omegaconf>=2.3",              # YAML config composition (used by Hydra)
    "hydra-core>=1.3",             # hierarchical config management

    # ── Experiment tracking ──────────────────────────────────────────
    "wandb>=0.16",

    # ── Hyperparameter search ────────────────────────────────────────
    "optuna>=3.5",
    "optuna-dashboard>=0.15",      # browser UI for Optuna studies

    # ── Data & analysis ──────────────────────────────────────────────
    "pandas>=2.0",
    "scikit-learn>=1.3",           # metrics, stratified splits
    "PyWavelets>=1.5",             # DWT preprocessing (Li et al. (CNN-Informer); Dokare & Gupta (DWT+SVM))
    "xgboost>=2.0",                # Classical ML baseline (Dash et al. (TF-Wearable); Djemal et al. (XGBoost-CS))
    "lightgbm>=4.0",               # Classical ML baseline (Dash et al. (TF-Wearable))

    # ── Visualization ────────────────────────────────────────────────
    "matplotlib>=3.8",
    "seaborn>=0.13",

    # ── Utilities ────────────────────────────────────────────────────
    "tqdm>=4.66",
    "rich>=13.0",                  # pretty console output
    "python-dotenv>=1.0",          # .env loading
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "pre-commit>=3.6",
    "ipykernel>=6.29",
    "jupyterlab>=4.0",
]
kaggle = [
    "kaggle>=1.6",                 # Kaggle API for dataset downloads
]
modelarts = [
    "moxing-framework",            # Huawei ModelArts data I/O
]

[project.scripts]
auras-train = "auras.training.trainer:main"
auras-search = "auras.experiment.optuna_search:main"
auras-export = "auras.deployment.converter:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 100
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
